Implement the Offloader traits for both prefixes docids types
