+++ crates/milli/src/update/new/offloader.rs
@@ -1,18 +1,14 @@
 use std::fs::File;
-use std::io::{self, BufReader, BufWriter, ErrorKind, Read as _, Seek as _, Write as _};
-
-use roaring::RoaringBitmap;
-
-use crate::CboRoaringBitmapCodec;
+use std::io::{self, BufReader, BufWriter, Seek as _};
 
 /// A data-structure that offloads prefixes and their serialized lengths to a file.
-pub struct PrefixIntegersOffloader<E, D> {
+pub struct Offloader<E, D> {
     file: BufWriter<File>,
     tmp_buffer: Vec<u8>,
     _marker: std::marker::PhantomData<(E, D)>,
 }
 
-impl<E, D> PrefixIntegersOffloader<E, D> {
+impl<E, D> Offloader<E, D> {
     pub fn new(file: File) -> Self {
         Self {
             file: BufWriter::new(file),
@@ -23,16 +19,16 @@ impl<E, D> PrefixIntegersOffloader<E, D> {
 
     pub fn push(&mut self, entry: E) -> io::Result<()>
     where
-        E: Encode,
+        E: Encoder,
     {
         self.tmp_buffer.clear();
         entry.encode(&mut self.tmp_buffer, &mut self.file)
     }
 
-    pub fn finish(self) -> io::Result<PrefixIntegersReader<D>> {
+    pub fn finish(self) -> io::Result<OffloadedReader<D>> {
         self.file.into_inner().map_err(|e| e.into_error()).and_then(|mut file| {
             file.rewind()?;
-            Ok(PrefixIntegersReader {
+            Ok(OffloadedReader {
                 file: BufReader::new(file),
                 first_tmp_buffer: Default::default(),
                 second_tmp_buffer: Default::default(),
@@ -42,17 +38,17 @@ impl<E, D> PrefixIntegersOffloader<E, D> {
     }
 }
 
-pub struct PrefixIntegersReader<D> {
+pub struct OffloadedReader<D> {
     file: BufReader<File>,
     first_tmp_buffer: Vec<u8>,
     second_tmp_buffer: Vec<u8>,
     _marker: std::marker::PhantomData<D>,
 }
 
-impl<D> PrefixIntegersReader<D> {
+impl<D> OffloadedReader<D> {
     pub fn next_entry<'a>(&'a mut self) -> io::Result<Option<D::Decoded>>
     where
-        D: Decode<'a>,
+        D: Decoder<'a>,
     {
         self.first_tmp_buffer.clear();
         self.second_tmp_buffer.clear();
@@ -60,11 +56,11 @@ impl<D> PrefixIntegersReader<D> {
     }
 }
 
-pub trait Encode {
+pub trait Encoder {
     fn encode<W: io::Write>(self, tmp_buffer: &mut Vec<u8>, writer: &mut W) -> io::Result<()>;
 }
 
-pub trait Decode<'b>: Sized {
+pub trait Decoder<'b>: Sized {
     type Decoded: 'b;
 
     fn decode<R: io::Read>(
@@ -72,92 +68,4 @@ pub trait Decode<'b>: Sized {
         second_tmp_buffer: &'b mut Vec<u8>,
         reader: &mut R,
     ) -> io::Result<Option<Self::Decoded>>;
-}
-
-/// Represents a prefix, its position in the field and the length the bitmap takes on disk.
-pub struct InPrefixIntegerEntry<'a> {
-    pub prefix: &'a str,
-    pub pos: u16,
-    pub bitmap: Option<RoaringBitmap>,
-}
-
-impl Encode for InPrefixIntegerEntry<'_> {
-    fn encode<W: io::Write>(self, tmp_buffer: &mut Vec<u8>, writer: &mut W) -> io::Result<()> {
-        let InPrefixIntegerEntry { prefix, pos, bitmap } = self;
-
-        // prefix length and prefix
-        let prefix_length: u8 =
-            prefix.len().try_into().map_err(|_| io::Error::other("prefix length too long"))?;
-        writer.write_all(bytemuck::bytes_of(&prefix_length))?;
-        writer.write_all(prefix.as_bytes())?;
-
-        // pos
-        writer.write_all(bytemuck::bytes_of(&pos))?;
-
-        // bitmap length and bitmap
-        let serialized_bytes = match bitmap {
-            Some(bitmap) => {
-                tmp_buffer.clear();
-                CboRoaringBitmapCodec::serialize_into_vec(&bitmap, tmp_buffer);
-                &tmp_buffer[..]
-            }
-            None => &[][..],
-        };
-        let serialized_bitmap_length: u32 = serialized_bytes
-            .len()
-            .try_into()
-            .map_err(|_| io::Error::other("serialized bitmap length too long"))?;
-        writer.write_all(bytemuck::bytes_of(&serialized_bitmap_length))?;
-        writer.write_all(serialized_bytes)?;
-
-        Ok(())
-    }
-}
-
-/// Represents a prefix, its position in the field and the length the bitmap takes on disk.
-pub struct OutPrefixIntegerEntry<'b> {
-    pub prefix: &'b str,
-    pub pos: u16,
-    pub bitmap: Option<&'b [u8]>,
-}
-
-pub struct OutPrefixIntegerEntryCodec;
-
-impl<'b> Decode<'b> for OutPrefixIntegerEntryCodec {
-    type Decoded = OutPrefixIntegerEntry<'b>;
-
-    fn decode<R: io::Read>(
-        first_tmp_buffer: &'b mut Vec<u8>,
-        second_tmp_buffer: &'b mut Vec<u8>,
-        reader: &mut R,
-    ) -> io::Result<Option<Self::Decoded>> {
-        // prefix length and prefix
-        let mut prefix_length: u16 = 0;
-        match reader.read_exact(bytemuck::bytes_of_mut(&mut prefix_length)) {
-            Ok(()) => (),
-            Err(e) if e.kind() == ErrorKind::UnexpectedEof => return Ok(None),
-            Err(e) => return Err(e),
-        }
-        first_tmp_buffer.resize(prefix_length as usize, 0);
-        reader.read_exact(first_tmp_buffer)?;
-        let prefix = std::str::from_utf8(first_tmp_buffer)
-            .map_err(|e| io::Error::new(ErrorKind::InvalidData, e))?;
-
-        // pos
-        let mut pos: u16 = 0;
-        reader.read_exact(bytemuck::bytes_of_mut(&mut pos))?;
-
-        // bitmap length and bitmap (bytes)
-        let mut bitmap_length: u16 = 0;
-        reader.read_exact(bytemuck::bytes_of_mut(&mut bitmap_length))?;
-        let bitmap = if bitmap_length == 0 {
-            None
-        } else {
-            second_tmp_buffer.resize(bitmap_length as usize, 0);
-            reader.read_exact(second_tmp_buffer)?;
-            Some(second_tmp_buffer.as_slice())
-        };
-
-        Ok(Some(Self::Decoded { prefix, pos, bitmap }))
-    }
 }
