+++ crates/milli/src/update/new/words_prefix_docids.rs
@@ -1,5 +1,5 @@
 use std::collections::BTreeSet;
-use std::io::{self, BufReader, BufWriter, ErrorKind, Read, Seek, Write};
+use std::io::{self, ErrorKind};
 use std::iter;
 
 use hashbrown::HashMap;
@@ -8,9 +8,8 @@ use heed::{Database, RwTxn};
 use rayon::iter::{IndexedParallelIterator as _, IntoParallelIterator, ParallelIterator as _};
 use roaring::{MultiOps, RoaringBitmap};
 
-use super::offloader::{InPrefixIntegerEntry, OutPrefixIntegerEntry, PrefixIntegersOffloader};
+use super::offloader::{Decoder, Encoder, Offloader};
 use crate::heed_codec::StrBEU16Codec;
-use crate::update::new::offloader::OutPrefixIntegerEntryCodec;
 use crate::{CboRoaringBitmapCodec, Index, Prefix, Result};
 
 struct WordPrefixDocids<'i> {
@@ -53,11 +52,9 @@ impl<'i> WordPrefixDocids<'i> {
             .into_par_iter()
             .enumerate()
             .map(|(thread_id, rtxn)| {
-                // `indexes` represent offsets at which prefixes computations were stored in the `file`.
-                let mut indexes = Vec::new();
-                let mut file = tempfile::tempfile().map(BufWriter::new)?;
+                let mut entries =
+                    tempfile::tempfile().map(Offloader::<_, OutPrefixEntryCodec>::new)?;
 
-                let mut buffer = Vec::new();
                 for (prefix_index, prefix) in prefix_to_compute.iter().enumerate() {
                     // Is prefix for another thread?
                     if prefix_index % thread_count != thread_id {
@@ -70,32 +67,18 @@ impl<'i> WordPrefixDocids<'i> {
                         .remap_types::<Str, CboRoaringBitmapCodec>()
                         .map(|result| result.map(|(_word, bitmap)| bitmap))
                         .union()?;
-
-                    buffer.clear();
-                    CboRoaringBitmapCodec::serialize_into_vec(&output, &mut buffer);
-                    indexes.push(PrefixEntry { prefix, serialized_length: buffer.len() });
-                    file.write_all(&buffer)?;
+                    entries.push(InPrefixEntry { prefix, bitmap: output })?;
                 }
 
-                Ok((indexes, file))
+                entries.finish().map_err(Into::into)
             })
             .collect::<Result<Vec<_>>>()?;
 
         // We iterate over all the collected and serialized bitmaps through
         // the files and entries to eventually put them in the final database.
-        let mut buffer = Vec::new();
-        for (index, file) in outputs {
-            let mut file = file.into_inner().map_err(|e| e.into_error())?;
-            file.rewind()?;
-            let mut file = BufReader::new(file);
-            for PrefixEntry { prefix, serialized_length } in index {
-                buffer.resize(serialized_length, 0);
-                file.read_exact(&mut buffer)?;
-                self.prefix_database.remap_data_type::<Bytes>().put(
-                    wtxn,
-                    prefix.as_bytes(),
-                    &buffer,
-                )?;
+        for mut entries in outputs {
+            while let Some(OutPrefixEntry { key, value }) = entries.next_entry()? {
+                self.prefix_database.remap_data_type::<Bytes>().put(wtxn, key, value)?;
             }
         }
 
@@ -103,10 +86,76 @@ impl<'i> WordPrefixDocids<'i> {
     }
 }
 
-/// Represents a prefix and the lenght the bitmap takes on disk.
-struct PrefixEntry<'a> {
-    prefix: &'a str,
-    serialized_length: usize,
+/// Represents a prefix, its position in the field and the length the bitmap takes on disk.
+pub struct InPrefixEntry<'a> {
+    pub prefix: &'a str,
+    pub bitmap: RoaringBitmap,
+}
+
+impl Encoder for InPrefixEntry<'_> {
+    fn encode<W: io::Write>(self, tmp_buffer: &mut Vec<u8>, writer: &mut W) -> io::Result<()> {
+        let InPrefixEntry { prefix, bitmap } = self;
+
+        // prefix length and prefix
+        let prefix_length: u8 =
+            prefix.len().try_into().map_err(|_| io::Error::other("prefix length too long"))?;
+        writer.write_all(bytemuck::bytes_of(&prefix_length))?;
+        writer.write_all(prefix.as_bytes())?;
+
+        // bitmap length and bitmap
+        let serialized_bytes = {
+            tmp_buffer.clear();
+            CboRoaringBitmapCodec::serialize_into_vec(&bitmap, tmp_buffer);
+            &tmp_buffer[..]
+        };
+        let serialized_bitmap_length: u32 = serialized_bytes
+            .len()
+            .try_into()
+            .map_err(|_| io::Error::other("serialized bitmap length too long"))?;
+        writer.write_all(bytemuck::bytes_of(&serialized_bitmap_length))?;
+        writer.write_all(serialized_bytes)?;
+
+        Ok(())
+    }
+}
+
+pub struct OutPrefixEntryCodec;
+
+/// Represents the key and value of a prefix integer entry.
+pub struct OutPrefixEntry<'b> {
+    pub key: &'b [u8],
+    pub value: &'b [u8],
+}
+
+impl<'b> Decoder<'b> for OutPrefixEntryCodec {
+    type Decoded = OutPrefixEntry<'b>;
+
+    fn decode<R: io::Read>(
+        first_tmp_buffer: &'b mut Vec<u8>,
+        second_tmp_buffer: &'b mut Vec<u8>,
+        reader: &mut R,
+    ) -> io::Result<Option<Self::Decoded>> {
+        // prefix length and prefix
+        let mut prefix_length: u16 = 0;
+        match reader.read_exact(bytemuck::bytes_of_mut(&mut prefix_length)) {
+            Ok(()) => (),
+            Err(e) if e.kind() == ErrorKind::UnexpectedEof => return Ok(None),
+            Err(e) => return Err(e),
+        }
+        first_tmp_buffer.resize(prefix_length as usize, 0);
+        reader.read_exact(first_tmp_buffer)?;
+
+        // bitmap length and bitmap (bytes)
+        let mut bitmap_length: u16 = 0;
+        reader.read_exact(bytemuck::bytes_of_mut(&mut bitmap_length))?;
+        second_tmp_buffer.resize(bitmap_length as usize, 0);
+        reader.read_exact(second_tmp_buffer)?;
+
+        Ok(Some(Self::Decoded {
+            key: first_tmp_buffer.as_slice(),
+            value: second_tmp_buffer.as_slice(),
+        }))
+    }
 }
 
 struct WordPrefixIntegerDocids<'i> {
@@ -154,8 +203,8 @@ impl<'i> WordPrefixIntegerDocids<'i> {
             .into_par_iter()
             .enumerate()
             .map(|(thread_id, rtxn)| {
-                let mut entries = tempfile::tempfile()
-                    .map(PrefixIntegersOffloader::<_, OutPrefixIntegerEntryCodec>::new)?;
+                let mut entries =
+                    tempfile::tempfile().map(Offloader::<_, OutPrefixIntegerEntryCodec>::new)?;
                 for (prefix_index, prefix) in prefixes.iter().enumerate() {
                     // Is prefix for another thread?
                     if prefix_index % thread_count != thread_id {
@@ -208,32 +257,111 @@ impl<'i> WordPrefixIntegerDocids<'i> {
 
         // We iterate over all the collected and serialized bitmaps through
         // the files and entries to eventually put them in the final database.
-        let mut key_buffer = Vec::new();
         for mut entries in outputs {
-            while let Some(entry) = entries.next_entry()? {
-                // TODO do the key_buffering in the OutTruc
-                key_buffer.clear();
-                key_buffer.extend_from_slice(entry.prefix.as_bytes());
-                key_buffer.push(0);
-                key_buffer.extend_from_slice(&entry.pos.to_be_bytes());
-
-                match entry.bitmap {
+            while let Some(OutPrefixIntegerEntry { key, value }) = entries.next_entry()? {
+                match value {
                     Some(bitmap_bytes) => {
                         self.prefix_database.remap_data_type::<Bytes>().put(
                             wtxn,
-                            &key_buffer,
+                            &key,
                             &bitmap_bytes,
                         )?;
                     }
                     None => {
-                        self.prefix_database.delete(wtxn, &key_buffer)?;
+                        self.prefix_database.delete(wtxn, &key)?;
                     }
                 }
-                drop(entry);
             }
         }
 
         Ok(())
+    }
+}
+
+/// Represents a prefix, its position in the field and the length the bitmap takes on disk.
+pub struct InPrefixIntegerEntry<'a> {
+    pub prefix: &'a str,
+    pub pos: u16,
+    pub bitmap: Option<RoaringBitmap>,
+}
+
+impl Encoder for InPrefixIntegerEntry<'_> {
+    fn encode<W: io::Write>(self, tmp_buffer: &mut Vec<u8>, writer: &mut W) -> io::Result<()> {
+        let InPrefixIntegerEntry { prefix, pos, bitmap } = self;
+
+        // prefix length and prefix
+        let prefix_length: u8 =
+            prefix.len().try_into().map_err(|_| io::Error::other("prefix length too long"))?;
+        writer.write_all(bytemuck::bytes_of(&prefix_length))?;
+        writer.write_all(prefix.as_bytes())?;
+
+        // pos
+        writer.write_all(bytemuck::bytes_of(&pos))?;
+
+        // bitmap length and bitmap
+        let serialized_bytes = match bitmap {
+            Some(bitmap) => {
+                tmp_buffer.clear();
+                CboRoaringBitmapCodec::serialize_into_vec(&bitmap, tmp_buffer);
+                &tmp_buffer[..]
+            }
+            None => &[][..],
+        };
+        let serialized_bitmap_length: u32 = serialized_bytes
+            .len()
+            .try_into()
+            .map_err(|_| io::Error::other("serialized bitmap length too long"))?;
+        writer.write_all(bytemuck::bytes_of(&serialized_bitmap_length))?;
+        writer.write_all(serialized_bytes)?;
+
+        Ok(())
+    }
+}
+
+pub struct OutPrefixIntegerEntryCodec;
+
+/// Represents the key and value of a prefix integer entry.
+pub struct OutPrefixIntegerEntry<'b> {
+    pub key: &'b [u8],
+    pub value: Option<&'b [u8]>,
+}
+
+impl<'b> Decoder<'b> for OutPrefixIntegerEntryCodec {
+    type Decoded = OutPrefixIntegerEntry<'b>;
+
+    fn decode<R: io::Read>(
+        first_tmp_buffer: &'b mut Vec<u8>,
+        second_tmp_buffer: &'b mut Vec<u8>,
+        reader: &mut R,
+    ) -> io::Result<Option<Self::Decoded>> {
+        // prefix length and prefix
+        let mut prefix_length: u16 = 0;
+        match reader.read_exact(bytemuck::bytes_of_mut(&mut prefix_length)) {
+            Ok(()) => (),
+            Err(e) if e.kind() == ErrorKind::UnexpectedEof => return Ok(None),
+            Err(e) => return Err(e),
+        }
+        first_tmp_buffer.resize(prefix_length as usize, 0);
+        reader.read_exact(first_tmp_buffer)?;
+        first_tmp_buffer.push(0);
+
+        // pos
+        let mut pos: u16 = 0;
+        reader.read_exact(bytemuck::bytes_of_mut(&mut pos))?;
+        first_tmp_buffer.extend_from_slice(&pos.to_be_bytes());
+
+        // bitmap length and bitmap (bytes)
+        let mut bitmap_length: u16 = 0;
+        reader.read_exact(bytemuck::bytes_of_mut(&mut bitmap_length))?;
+        let bitmap = if bitmap_length == 0 {
+            None
+        } else {
+            second_tmp_buffer.resize(bitmap_length as usize, 0);
+            reader.read_exact(second_tmp_buffer)?;
+            Some(second_tmp_buffer.as_slice())
+        };
+
+        Ok(Some(Self::Decoded { key: first_tmp_buffer.as_slice(), value: bitmap }))
     }
 }
